refactor(utils): move findTsConfig to its own module
The findTsConfig function was moved from find-files.ts to a new module
tsconfig.ts. This change improves the organization of the code by
separating concerns. Each module now has a single responsibility,
making the code easier to understand and maintain.
